# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/chat/project/{projectId}/participant/{participantId}": {
      "post": {
        "description": "Creates a new chat or returns existing one between admin and participant",
        "operationId": "ChatController_createOrGetChat",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID of the project",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "participantId",
            "required": true,
            "in": "path",
            "description": "ID of the participant",
            "schema": {
              "example": 2,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved or created chat",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "User is not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create or get chat with participant",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/chat/{chatId}/messages": {
      "get": {
        "description": "Retrieves a paginated list of messages for a specific chat",
        "operationId": "ChatController_getMessages",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "ID of the chat",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of messages to return (default: 20)",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of messages to skip (default: 0)",
            "schema": {
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved messages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedMessagesResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "User is not authenticated"
          },
          "403": {
            "description": "User does not have access to this chat"
          },
          "404": {
            "description": "Chat not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get messages for a chat",
        "tags": [
          "chat"
        ]
      },
      "post": {
        "description": "Creates a new message in a specific chat",
        "operationId": "ChatController_createMessage",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "ID of the chat",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageViaApiDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid message content or spam protection triggered"
          },
          "401": {
            "description": "User is not authenticated"
          },
          "403": {
            "description": "User does not have access to this chat"
          },
          "404": {
            "description": "Chat not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send a message to a chat",
        "tags": [
          "chat"
        ]
      }
    },
...
    "/api/mobile-participants": {
      "post": {
        "description": "Creates a new participant for a specific project with external participant ID. No authentication required for mobile app usage.",
        "operationId": "MobileParticipantController_createParticipant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectParticipantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created participant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectParticipantResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid participant data or participant ID already exists"
          }
        },
        "summary": "Create a new project participant (Mobile App)",
        "tags": [
          "mobile-participants"
        ]
      }
    },
    "/api/mobile-chat/project/{projectId}/participant/{participantId}": {
      "post": {
        "description": "Gets or creates a chat between admin and participant. No authentication required for mobile app usage.",
        "operationId": "MobileChatController_getOrCreateChat",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID of the project",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "participantId",
            "required": true,
            "in": "path",
            "description": "External participant ID (string) provided by mobile app",
            "schema": {
              "example": "mobile_user_123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved or created chat",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project or participant data"
          }
        },
        "summary": "Get or create chat (Mobile App)",
        "tags": [
          "mobile-chat"
        ]
      }
    },
    "/api/mobile-chat/{chatId}/messages": {
      "get": {
        "description": "Retrieves a paginated list of messages for a specific chat. No authentication required for mobile app usage.",
        "operationId": "MobileChatController_getMessages",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "ID of the chat",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "participantId",
            "required": true,
            "in": "query",
            "description": "External participant ID (string) provided by mobile app",
            "schema": {
              "example": "mobile_user_123",
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "query",
            "description": "ID of the project",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of messages to return (default: 20)",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of messages to skip (default: 0)",
            "schema": {
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved messages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedMessagesResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Chat not found"
          }
        },
        "summary": "Get messages for a chat (Mobile App)",
        "tags": [
          "mobile-chat"
        ]
      },
      "post": {
        "description": "Creates a new message in a specific chat. No authentication required for mobile app usage.",
        "operationId": "MobileChatController_createMessage",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "ID of the chat",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "participantId",
            "required": true,
            "in": "query",
            "description": "External participant ID (string) provided by mobile app",
            "schema": {
              "example": "mobile_user_123",
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "query",
            "description": "ID of the project",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageViaApiDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid message content or spam protection triggered"
          },
          "404": {
            "description": "Chat not found"
          }
        },
        "summary": "Send a message to a chat (Mobile App)",
        "tags": [
          "mobile-chat"
        ]
      }
    }
  },
  "info": {
    "title": "Chat CRM API",
    "description": "API documentation for Chat CRM application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "ProjectDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Project ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Project name",
            "example": "My Project"
          },
          "uniqueId": {
            "type": "string",
            "description": "Project unique ID",
            "example": "proj123"
          }
        },
        "required": [
          "id",
          "name",
          "uniqueId"
        ]
      },
      "ChatParticipantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "User ID",
            "example": 1
          },
          "email": {
            "type": "object",
            "description": "User email (optional for participants)",
            "example": "user@example.com",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "description": "User first name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "User last name",
            "example": "Doe"
          },
          "uniqueId": {
            "type": "string",
            "description": "User unique ID (for participants)",
            "example": "user123"
          },
          "isOnline": {
            "type": "boolean",
            "description": "Whether the user is online",
            "example": false
          },
          "lastSeen": {
            "type": "object",
            "description": "Last seen timestamp",
            "example": "2025-01-15T10:00:00.000Z",
            "nullable": true
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "isOnline",
          "lastSeen"
        ]
      },
      "LastMessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Message ID",
            "example": 1
          },
          "content": {
            "type": "string",
            "description": "Message content",
            "example": "Hello! How can I help you?"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Message creation timestamp",
            "example": "2025-01-15T10:30:00.000Z"
          },
          "senderId": {
            "type": "number",
            "description": "ID of the user who sent the message",
            "example": 1
          }
        },
        "required": [
          "id",
          "content",
          "createdAt",
          "senderId"
        ]
      },
      "ChatResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Chat ID",
            "example": 1
          },
          "projectId": {
            "type": "number",
            "description": "Project ID",
            "example": 1
          },
          "project": {
            "description": "Project information",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProjectDto"
              }
            ]
          },
          "admin": {
            "description": "Admin user information",
            "allOf": [
              {
                "$ref": "#/components/schemas/ChatParticipantDto"
              }
            ]
          },
          "participant": {
            "description": "Participant user information",
            "allOf": [
              {
                "$ref": "#/components/schemas/ChatParticipantDto"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Chat creation timestamp",
            "example": "2025-01-15T10:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Chat last update timestamp",
            "example": "2025-01-15T10:30:00.000Z"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the chat is active",
            "example": true
          },
          "lastMessage": {
            "description": "Last message in the chat",
            "allOf": [
              {
                "$ref": "#/components/schemas/LastMessageDto"
              }
            ]
          },
          "unreadCount": {
            "type": "number",
            "description": "Number of unread messages for the requesting user",
            "example": 3
          },
          "lastMessageAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of the last message",
            "example": "2025-01-15T10:30:00.000Z"
          }
        },
        "required": [
          "id",
          "projectId",
          "project",
          "admin",
          "participant",
          "createdAt",
          "updatedAt",
          "isActive",
          "unreadCount"
        ]
      },
      "MessageSenderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "User ID",
            "example": 1
          },
          "email": {
            "type": "object",
            "description": "User email",
            "example": "user@example.com",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "description": "User first name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "User last name",
            "example": "Doe"
          },
          "uniqueId": {
            "type": "string",
            "description": "User unique identifier (for participants)",
            "example": "user123"
          },
          "isOnline": {
            "type": "boolean",
            "description": "Whether the user is currently online",
            "example": true
          },
          "lastSeen": {
            "type": "object",
            "description": "Last time the user was seen",
            "example": "2025-01-15T10:30:00.000Z",
            "nullable": true
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "isOnline"
        ]
      },
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Message ID",
            "example": 1
          },
          "chatId": {
            "type": "number",
            "description": "Chat ID where the message belongs",
            "example": 1
          },
          "senderId": {
            "type": "number",
            "description": "ID of the user who sent the message",
            "example": 1
          },
          "content": {
            "type": "string",
            "description": "Message content",
            "example": "Hello! How can I help you today?"
          },
          "type": {
            "type": "string",
            "description": "Type of the message. IMAGE for image files, FILE for other files, TEXT for text messages",
            "enum": [
              "TEXT",
              "EMOJI",
              "FILE",
              "IMAGE",
              "SYSTEM"
            ],
            "example": "TEXT"
          },
          "metadata": {
            "type": "object",
            "description": "Additional metadata for the message",
            "example": {
              "fileUrl": "https://storage.googleapis.com/bucket/chat/1/uuid.jpg",
              "fileName": "image.jpg",
              "fileSize": 1024000,
              "mimeType": "image/jpeg",
              "originalName": "photo.jpg"
            }
          },
          "read": {
            "type": "boolean",
            "description": "Whether the message has been read",
            "example": false
          },
          "readAt": {
            "type": "object",
            "description": "Timestamp when the message was read",
            "example": "2025-01-15T10:35:00.000Z",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Message creation timestamp",
            "example": "2025-01-15T10:30:00.000Z"
          },
          "sender": {
            "description": "Information about the message sender",
            "allOf": [
              {
                "$ref": "#/components/schemas/MessageSenderDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "chatId",
          "senderId",
          "content",
          "type",
          "read",
          "createdAt",
          "sender"
        ]
      },
      "PaginatedMessagesResponseDto": {
        "type": "object",
        "properties": {
          "messages": {
            "description": "Array of messages",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageResponseDto"
            }
          },
          "totalCount": {
            "type": "number",
            "description": "Total number of messages in the chat",
            "example": 50
          },
          "limit": {
            "type": "number",
            "description": "Number of messages per page",
            "example": 20
          },
          "offset": {
            "type": "number",
            "description": "Number of messages skipped",
            "example": 0
          }
        },
        "required": [
          "messages",
          "totalCount",
          "limit",
          "offset"
        ]
      },
      "CreateMessageViaApiDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "Content of the message",
            "example": "Hello! How can I help you today?",
            "maxLength": 1000
          },
          "type": {
            "type": "string",
            "description": "Type of the message. Use IMAGE for image files, FILE for other files, TEXT for text messages",
            "enum": [
              "TEXT",
              "EMOJI",
              "FILE",
              "IMAGE",
              "SYSTEM"
            ],
            "example": "TEXT",
            "default": "TEXT"
          },
          "metadata": {
            "type": "object",
            "description": "Additional metadata for the message (e.g., file info, emoji data)",
            "example": {
              "fileUrl": "https://storage.googleapis.com/bucket/chat/1/uuid.jpg",
              "fileName": "image.jpg",
              "fileSize": 1024000,
              "mimeType": "image/jpeg",
              "originalName": "photo.jpg"
            }
          }
        },
        "required": [
          "content"
        ]
      },
      "CreateProjectParticipantDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "number",
            "description": "ID of the project where the participant will be added",
            "example": 1,
            "minimum": 1
          },
          "participantId": {
            "type": "string",
            "description": "External participant ID provided by mobile app",
            "example": "mobile_user_123",
            "minLength": 3,
            "maxLength": 50
          },
          "firstName": {
            "type": "string",
            "description": "First name of the participant (optional for mobile app)",
            "example": "John",
            "maxLength": 50
          },
          "lastName": {
            "type": "string",
            "description": "Last name of the participant (optional for mobile app)",
            "example": "Doe",
            "maxLength": 50
          },
          "email": {
            "type": "string",
            "description": "Email address of the participant (optional for mobile app)",
            "example": "john.doe@example.com"
          }
        },
        "required": [
          "projectId",
          "participantId"
        ]
      },
      "ProjectParticipantResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Participant ID",
            "example": 1
          },
          "projectId": {
            "type": "number",
            "description": "Project ID",
            "example": 1
          },
          "userId": {
            "type": "number",
            "description": "User ID",
            "example": 2
          },
          "participantId": {
            "type": "string",
            "description": "External participant ID provided by mobile app",
            "example": "mobile_user_123"
          },
          "firstName": {
            "type": "string",
            "description": "First name of the participant",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "Last name of the participant",
            "example": "Doe"
          },
          "email": {
            "type": "object",
            "description": "Email address of the participant (optional)",
            "example": "john.doe@example.com",
            "nullable": true
          },
          "isOnline": {
            "type": "boolean",
            "description": "Whether the participant is online",
            "example": false
          },
          "lastSeen": {
            "type": "object",
            "description": "Last seen timestamp",
            "example": "2025-01-15T10:00:00.000Z",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the participant was added to the project",
            "example": "2025-01-15T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "projectId",
          "userId",
          "participantId",
          "firstName",
          "lastName",
          "email",
          "isOnline",
          "lastSeen",
          "createdAt"
        ]
      }
    }
  }
}